--- src/pages/Home/components/loader/customLoader.module.css
@@ -1,313 +1,0 @@
-.container {
-  background-color: white;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-  z-index: 9999;
-}
-
-.text {
-  font: var(--font-h1);
-  color: var(--text-main-color, #141414);
-  margin-left: 10px;
-}
-
-.loader {
-  width: 60px;
-  height: 60px;
-  position: relative;
-  transform-style: preserve-3d;
-  transform: rotateX(45deg) rotateY(45deg);
-}
-
-/* Внешний куб */
-.cube {
-  position: absolute;
-  width: 60px;
-  height: 60px;
-  transform-style: preserve-3d;
-}
-
-/* Грани внешнего куба */
-.face {
-  position: absolute;
-  width: 60px;
-  height: 60px;
-  background: #5b93f4;
-  border: 1px solid rgba(255, 255, 255, 0.3);
-  box-sizing: border-box;
-  opacity: 0.8;
-}
-
-/* Позиционирование граней внешнего куба */
-.face:nth-child(1) {
-  transform: translateZ(30px); /* Передняя грань */
-}
-.face:nth-child(2) {
-  transform: translateZ(-30px); /* Задняя грань */
-}
-.face:nth-child(3) {
-  transform: translateX(30px) rotateY(90deg); /* Правая грань */
-}
-.face:nth-child(4) {
-  transform: translateX(-30px) rotateY(90deg); /* Левая грань */
-}
-.face:nth-child(5) {
-  transform: translateY(30px) rotateX(90deg); /* Верхняя грань */
-}
-.face:nth-child(6) {
-  transform: translateY(-30px) rotateX(90deg); /* Нижняя грань */
-}
-
-/* Внутренние кубы */
-.innerCube {
-  position: absolute;
-  width: 15px;
-  height: 15px;
-  transform-style: preserve-3d;
-  animation: pulse 3s infinite ease-in-out; /* Изменено на infinite ease-in-out */
-  /* Центрируем внутренние кубы */
-  top: 50%;
-  left: 50%;
-  margin-left: -7.5px;
-  margin-top: -7.5px;
-}
-
-/* Позиции внутренних кубов относительно центра внешнего куба */
-.innerCube:nth-child(1) {
-  transform: translate3d(-22.5px, -22.5px, -22.5px);
-}
-.innerCube:nth-child(2) {
-  transform: translate3d(0px, -22.5px, -22.5px);
-}
-.innerCube:nth-child(3) {
-  transform: translate3d(22.5px, -22.5px, -22.5px);
-}
-.innerCube:nth-child(4) {
-  transform: translate3d(-22.5px, 0px, -22.5px);
-}
-.innerCube:nth-child(5) {
-  transform: translate3d(22.5px, 0px, -22.5px);
-}
-.innerCube:nth-child(6) {
-  transform: translate3d(-22.5px, 22.5px, -22.5px);
-}
-.innerCube:nth-child(7) {
-  transform: translate3d(0px, 22.5px, -22.5px);
-}
-.innerCube:nth-child(8) {
-  transform: translate3d(22.5px, 22.5px, -22.5px);
-}
-
-.innerCube:nth-child(9) {
-  transform: translate3d(-22.5px, -22.5px, 0px);
-}
-.innerCube:nth-child(10) {
-  transform: translate3d(22.5px, -22.5px, 0px);
-}
-.innerCube:nth-child(11) {
-  transform: translate3d(-22.5px, 22.5px, 0px);
-}
-.innerCube:nth-child(12) {
-  transform: translate3d(22.5px, 22.5px, 0px);
-}
-
-.innerCube:nth-child(13) {
-  transform: translate3d(-22.5px, -22.5px, 22.5px);
-}
-.innerCube:nth-child(14) {
-  transform: translate3d(0px, -22.5px, 22.5px);
-}
-.innerCube:nth-child(15) {
-  transform: translate3d(22.5px, -22.5px, 22.5px);
-}
-.innerCube:nth-child(16) {
-  transform: translate3d(-22.5px, 0px, 22.5px);
-}
-.innerCube:nth-child(17) {
-  transform: translate3d(22.5px, 0px, 22.5px);
-}
-.innerCube:nth-child(18) {
-  transform: translate3d(-22.5px, 22.5px, 22.5px);
-}
-.innerCube:nth-child(19) {
-  transform: translate3d(0px, 22.5px, 22.5px);
-}
-.innerCube:nth-child(20) {
-  transform: translate3d(22.5px, 22.5px, 22.5px);
-}
-
-/* Грани внутренних кубов */
-.innerFace {
-  position: absolute;
-  width: 15px;
-  height: 15px;
-  background: #043b94;
-  border: 1px solid rgba(255, 255, 255, 0.5);
-  box-sizing: border-box;
-  transform-style: preserve-3d;
-}
-
-/* Позиционирование граней внутренних кубов */
-.innerFace:nth-child(1) {
-  transform: translateZ(7.5px);
-}
-.innerFace:nth-child(2) {
-  transform: translateZ(-7.5px);
-}
-.innerFace:nth-child(3) {
-  transform: translateX(7.5px) rotateY(90deg);
-}
-.innerFace:nth-child(4) {
-  transform: translateX(-7.5px) rotateY(90deg);
-}
-.innerFace:nth-child(5) {
-  transform: translateY(7.5px) rotateX(90deg);
-}
-.innerFace:nth-child(6) {
-  transform: translateY(-7.5px) rotateX(90deg);
-}
-
-/* Анимация пульсации для внутренних кубов - бесконечная */
-@keyframes pulse {
-  0%,
-  100% {
-    opacity: 0.4;
-    transform: translate3d(var(--tx, 0), var(--ty, 0), var(--tz, 0)) scale(0.8);
-  }
-  50% {
-    opacity: 1;
-    transform: translate3d(var(--tx, 0), var(--ty, 0), var(--tz, 0)) scale(1.1);
-  }
-}
-
-/* Добавляем задержки анимации для создания волнового эффекта */
-.innerCube:nth-child(1) {
-  --tx: -22.5px;
-  --ty: -22.5px;
-  --tz: -22.5px;
-  animation-delay: 0s;
-}
-.innerCube:nth-child(2) {
-  --tx: 0px;
-  --ty: -22.5px;
-  --tz: -22.5px;
-  animation-delay: 0.15s;
-}
-.innerCube:nth-child(3) {
-  --tx: 22.5px;
-  --ty: -22.5px;
-  --tz: -22.5px;
-  animation-delay: 0.3s;
-}
-.innerCube:nth-child(4) {
-  --tx: -22.5px;
-  --ty: 0px;
-  --tz: -22.5px;
-  animation-delay: 0.45s;
-}
-.innerCube:nth-child(5) {
-  --tx: 22.5px;
-  --ty: 0px;
-  --tz: -22.5px;
-  animation-delay: 0.6s;
-}
-.innerCube:nth-child(6) {
-  --tx: -22.5px;
-  --ty: 22.5px;
-  --tz: -22.5px;
-  animation-delay: 0.75s;
-}
-.innerCube:nth-child(7) {
-  --tx: 0px;
-  --ty: 22.5px;
-  --tz: -22.5px;
-  animation-delay: 0.9s;
-}
-.innerCube:nth-child(8) {
-  --tx: 22.5px;
-  --ty: 22.5px;
-  --tz: -22.5px;
-  animation-delay: 1.05s;
-}
-
-.innerCube:nth-child(9) {
-  --tx: -22.5px;
-  --ty: -22.5px;
-  --tz: 0px;
-  animation-delay: 1.2s;
-}
-.innerCube:nth-child(10) {
-  --tx: 22.5px;
-  --ty: -22.5px;
-  --tz: 0px;
-  animation-delay: 1.35s;
-}
-.innerCube:nth-child(11) {
-  --tx: -22.5px;
-  --ty: 22.5px;
-  --tz: 0px;
-  animation-delay: 1.5s;
-}
-.innerCube:nth-child(12) {
-  --tx: 22.5px;
-  --ty: 22.5px;
-  --tz: 0px;
-  animation-delay: 1.65s;
-}
-
-.innerCube:nth-child(13) {
-  --tx: -22.5px;
-  --ty: -22.5px;
-  --tz: 22.5px;
-  animation-delay: 1.8s;
-}
-.innerCube:nth-child(14) {
-  --tx: 0px;
-  --ty: -22.5px;
-  --tz: 22.5px;
-  animation-delay: 1.95s;
-}
-.innerCube:nth-child(15) {
-  --tx: 22.5px;
-  --ty: -22.5px;
-  --tz: 22.5px;
-  animation-delay: 2.1s;
-}
-.innerCube:nth-child(16) {
-  --tx: -22.5px;
-  --ty: 0px;
-  --tz: 22.5px;
-  animation-delay: 2.25s;
-}
-.innerCube:nth-child(17) {
-  --tx: 22.5px;
-  --ty: 0px;
-  --tz: 22.5px;
-  animation-delay: 2.4s;
-}
-.innerCube:nth-child(18) {
-  --tx: -22.5px;
-  --ty: 22.5px;
-  --tz: 22.5px;
-  animation-delay: 2.55s;
-}
-.innerCube:nth-child(19) {
-  --tx: 0px;
-  --ty: 22.5px;
-  --tz: 22.5px;
-  animation-delay: 2.7s;
-}
-.innerCube:nth-child(20) {
-  --tx: 22.5px;
-  --ty: 22.5px;
-  --tz: 22.5px;
-  animation-delay: 2.85s;
-}
