+++ src/pages/Home/components/loader/index.ts
@@ -1,4 +1,1 @@
 export { default as Loader } from './loader';
-export { default as Loader_2 } from './loader_2';
-export { default as Loader_3 } from './loader_3';
-export { default as CustomLoader } from './customLoader';
