+++ src/pages/Home/components/loader/loader.module.css
@@ -1,86 +1,114 @@
 .container {
-  background-color: white;
   position: absolute;
   top: 0;
   left: 0;
   width: 100vw;
   height: 100vh;
   display: flex;
+  flex-direction: row;
   justify-content: center;
   align-items: center;
+  gap: 5px;
   z-index: 9999;
-}
-.loader {
-  --s: 25px;
-  --g: 5px;
-
-  width: calc(2 * (1.353 * var(--s) + var(--g)));
-  aspect-ratio: 1;
-  background:
-    linear-gradient(var(--blue) 0 0) left/50% 100% no-repeat,
-    conic-gradient(
-      from -90deg at var(--s) calc(0.353 * var(--s)),
-      #fff 135deg,
-      #666 0 270deg,
-      #aaa 0
-    );
-  background-blend-mode: multiply;
-  --_m:
-    linear-gradient(
-      to bottom right,
-      #0000 calc(0.25 * var(--s)),
-      #000 0 calc(100% - calc(0.25 * var(--s)) - 1.414 * var(--g)),
-      #0000 0
-    ),
-    conic-gradient(
-      from -90deg at right var(--g) bottom var(--g),
-      #000 90deg,
-      #0000 0
-    );
-  -webkit-mask: var(--_m);
-  mask: var(--_m);
-  background-size: 50% 50%;
-  -webkit-mask-size: 50% 50%;
-  mask-size: 50% 50%;
-  -webkit-mask-composite: source-in;
-  mask-composite: intersect;
-  animation: l9 1.5s infinite;
-}
-@keyframes l9 {
-  0%,
-  12.5% {
-    background-position:
-      0% 0%,
-      0 0;
-  }
-  12.6%,
-  37.5% {
-    background-position:
-      100% 0%,
-      0 0;
-  }
-  37.6%,
-  62.5% {
-    background-position:
-      100% 100%,
-      0 0;
-  }
-  62.6%,
-  87.5% {
-    background-position:
-      0% 100%,
-      0 0;
-  }
-  87.6%,
-  100% {
-    background-position:
-      0% 0%,
-      0 0;
-  }
+  background-color: transparent;
 }
 
 .text {
   font: var(--font-h1);
-  color: var(--text-main-color, #141414);
+  color: var(--text-for-light-button);
   margin-left: 10px;
+}
+
+.dots::after {
+  content: '...';
+  display: inline-block;
+  opacity: 0;
+  animation: dotsSequence 1.5s steps(3, end) infinite;
+}
+
+@keyframes dotsSequence {
+  0% {
+    opacity: 0;
+  }
+  10% {
+    opacity: 1;
+  }
+  33.33% {
+    opacity: 1;
+  }
+  43.33% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 0;
+  }
+}
+
+.loader {
+  display: grid;
+  grid-template-columns: repeat(3, 20px);
+  grid-template-rows: repeat(3, 20px);
+  gap: 10px;
+  width: max-content;
+  height: max-content;
+  margin-bottom: 10px;
+}
+
+.cube {
+  width: 25px;
+  height: 25px;
+  border-radius: 5px;
+  background-color: #3498db; /* цвет */
+  animation: pulse 1.4s infinite ease-in-out;
+}
+
+@keyframes pulse {
+  0%,
+  100% {
+    opacity: 1;
+    transform: scale(1);
+    background-color: #3498db;
+  }
+  50% {
+    opacity: 0.5;
+    transform: scale(0.8);
+    background-color: #2ecc71;
+  }
+}
+
+.cube:nth-child(1) {
+  animation-delay: 0s;
+}
+.cube:nth-child(2) {
+  animation-delay: 0.2s;
+}
+.cube:nth-child(3) {
+  animation-delay: 0.4s;
+}
+.cube:nth-child(4) {
+  animation-delay: 0.2s;
+}
+.cube:nth-child(5) {
+  animation-delay: 0.4s;
+}
+.cube:nth-child(6) {
+  animation-delay: 0.6s;
+}
+.cube:nth-child(7) {
+  animation-delay: 0.4s;
+}
+.cube:nth-child(8) {
+  animation-delay: 0.6s;
+}
+.cube:nth-child(9) {
+  animation-delay: 0.8s;
+}
+
+@media screen and (max-width: 833px) {
+  .container {
+    flex-direction: column;
+  }
+  .text {
+    font: var(--font-h1-mobile);
+  }
 }
