+++ src/pages/Home/components/loader/loader.tsx
@@ -1,32 +1,22 @@
 import styles from './loader.module.css';
-import { useEffect } from 'react';
 
-export type LoaderProps = {
-  openLoader: boolean;
-  doClose: () => void;
-};
-
-const Loader = (props: LoaderProps) => {
-  const { openLoader, doClose } = props;
-
-  useEffect(() => {
-    if (openLoader) {
-      const timer = setTimeout(() => {
-        console.log('20 seconds passed, closing loader');
-        doClose();
-      }, 20000);
-      return () => clearTimeout(timer);
-    }
-  }, [openLoader, doClose]);
-
-  if (!openLoader) {
-    return null;
-  }
-
+const Loader = () => {
   return (
     <div className={styles.container}>
-      <div className={styles.loader}> </div>
-      <span className={styles.text}>Обработка данных</span>
+      <div className={styles.loader}>
+        <div className={styles.cube}></div>
+        <div className={styles.cube}></div>
+        <div className={styles.cube}></div>
+        <div className={styles.cube}></div>
+        <div className={styles.cube}></div>
+        <div className={styles.cube}></div>
+        <div className={styles.cube}></div>
+        <div className={styles.cube}></div>
+        <div className={styles.cube}></div>
+      </div>
+      <span className={styles.text}>
+        Формируется отчёт<span className={styles.dots}></span>
+      </span>
     </div>
   );
 };
