--- src/pages/Home/components/loader/loader_3.module.css
@@ -1,114 +1,0 @@
-.container {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  gap: 5px;
-  z-index: 9999;
-  background-color: transparent;
-}
-
-.text {
-  font: var(--font-h1);
-  color: var(--text-for-light-button);
-  margin-left: 10px;
-}
-
-.dots::after {
-  content: '...';
-  display: inline-block;
-  opacity: 0;
-  animation: dotsSequence 1.5s steps(3, end) infinite;
-}
-
-@keyframes dotsSequence {
-  0% {
-    opacity: 0;
-  }
-  10% {
-    opacity: 1;
-  }
-  33.33% {
-    opacity: 1;
-  }
-  43.33% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 0;
-  }
-}
-
-.loader {
-  display: grid;
-  grid-template-columns: repeat(3, 20px);
-  grid-template-rows: repeat(3, 20px);
-  gap: 10px;
-  width: max-content;
-  height: max-content;
-  margin-bottom: 10px;
-}
-
-.cube {
-  width: 25px;
-  height: 25px;
-  border-radius: 5px;
-  background-color: #3498db; /* цвет */
-  animation: pulse 1.4s infinite ease-in-out;
-}
-
-@keyframes pulse {
-  0%,
-  100% {
-    opacity: 1;
-    transform: scale(1);
-    background-color: #3498db;
-  }
-  50% {
-    opacity: 0.5;
-    transform: scale(0.8);
-    background-color: #2ecc71;
-  }
-}
-
-.cube:nth-child(1) {
-  animation-delay: 0s;
-}
-.cube:nth-child(2) {
-  animation-delay: 0.2s;
-}
-.cube:nth-child(3) {
-  animation-delay: 0.4s;
-}
-.cube:nth-child(4) {
-  animation-delay: 0.2s;
-}
-.cube:nth-child(5) {
-  animation-delay: 0.4s;
-}
-.cube:nth-child(6) {
-  animation-delay: 0.6s;
-}
-.cube:nth-child(7) {
-  animation-delay: 0.4s;
-}
-.cube:nth-child(8) {
-  animation-delay: 0.6s;
-}
-.cube:nth-child(9) {
-  animation-delay: 0.8s;
-}
-
-@media screen and (max-width: 833px) {
-  .container {
-    flex-direction: column;
-  }
-  .text {
-    font: var(--font-h1-mobile);
-  }
-}
